refactor: change impementation of bugfix
instead of using try/catch to catch errors when trying to deete nonexistent coverBlock, query DB to see if it exists before deleting

+++ apps/api-journeys/src/app/modules/block/image/image.resolver.ts
@@ -90,19 +90,15 @@ export class ImageBlockResolver {
       })
       // Delete old coverBlock
       if (parentBlock.coverBlockId != null) {
-        try {
+        const coverBlockExists = await this.blockService.validateBlock(
+          parentBlock.coverBlockId,
+          parentBlock.id
+        )
+        if (coverBlockExists) {
           await this.blockService.removeBlockAndChildren(
             parentBlock.coverBlockId,
             parentBlock.journeyId
           )
-        } catch (error) {
-          // ignore error thrown if old coverBlock was already deleted
-          if (
-            error.isArangoError !== true ||
-            error.response.body.errorMessage !== 'document not found'
-          ) {
-            throw new Error(error)
-          }
         }
       }
